Chambers.jl

We present a Julia package for computing the connected components in the complement of an arrangement of real hypersurfaces in affine or projective space.

Our input consists of $k$ polynomials in $n$ variables. The output is a list of all chambers (i.e. connected components) of the complement of the zero set of these polynomials.

For each chamber $C$ we report the Euler characteristic and well-chosen sample points. The list is grouped according to sign vectors $\sigma \in \{-1,+1 \}^k$, where $\sigma_i$ is the sign of the $i$th polynomial on $C$. For hyperplane arrangements, each chamber is uniquely characterized by its sign vector. However, in our situation, each sign vector $\sigma$ typically corresponds to multiple connected components.

Example: two concentric circles

Let us consider two concentric circles. For instance, we could take the two circles $f_1 = x^2 + y^2 - 1=0$ and $f_2=x^2 + y^2 - 4=0$ centered at the origin. To compute the chambers of $\mathcal{U} = \{ u \in \mathbb{R}^2 \mid f_1(u) \cdot f_2(u) \not= 0 \}$ we can use the following code:

julia> using Chambers
julia> @var x y;
julia> f_1 = x^2 + y^2 - 1;
julia> f_2 = x^2 + y^2 - 4;
julia> f = [f_1; f_2]
julia> C = chambers(f)
ChambersResult with 3 chambers:
=============================
9 complex critical points
9 real critical points
╭──────────────┬─────────────────────────────────╮
│ sign pattern │ chambers                         │
├──────────────┼─────────────────────────────────┤
│ - -          │ number = 1                      │
│              │ χ = 1, μ = [1, 0, 0], bounded   │
│ + -          │ number = 1                      │
│              │ χ = 0, μ = [2, 2, 0], bounded   │
│ + +          │ number = 1                      │
│              │ χ = 0, μ = [2, 2, 0], unbounded │
╰──────────────┴─────────────────────────────────╯

The output shows that $\mathcal U$ has three chambers. The first chamber has sign pattern $++$. This means that, on this chamber, both $f_1$ and $f_2$ are positive. On the second chamber, both $f_1$ and $f_2$ are negative, so it is the contractible chamber in the middle. The software correctly reports that this chamber has Euler characteristic 1. The other two chambers each have one hole and thus have Euler characteristic 0.

Example: 3 random polynomials in 4 variables

We can set up a random example as follows.

julia> using Chambers
julia> @var v[1:3];
julia> f = [rand_poly(Float64, v, d) for d in [2, 3, 3]];
julia> C = chambers(f)

Here, f consists of 3 random polynomials in v. The degrees of these polynomials are [2, 3, 3]. The coefficients are chosen from a Gaussian distribution.

Documentation: Output

Chambers.ChambersResult — Type

ChambersResult

A struct that collects all chambers in the complement of a hypersurface arragement.

Chambers.Chamber — Type

Chamber

A struct that contains all information about a chamber.

Documentation: Main functions

Chambers.chambers — Function

chambers(C::ChambersResult)

Returns the vector of chambers in C.

chambers(f::Vector{Expression})
chambers(f::System)

Input a list of hypersurfaces 'f = [f1,...fk]'. Outputs the chambers in the complement of the hypersurface arrangement, whether they are bounded or not, their sign patterns, Euler characteristic and the indices of the critical points in each chamber.

Options:

show_progress = true: if true, prints the progress of the computation to the terminal.

projective_fusion = true: if true, the algorithm computes which of the chambers are fused at infinity.

s: a list of integers [s_1, ..., s_k, s_{k+1}] such that s_1, ..., s_k>0, s_{k+1}<0 and 2 s_{k+1} > s_1 deg(f_1) + ... + s_k deg(f_k).

epsilon = 1e-6: how close from each critical point do we do the path tracking.

reltol = 1e-6, abstol = 1e-9: parameters for the accuracy of the ODE solver.

monodromy_options = MonodromyOptions(max_loops_no_progress = 25): pass options for monodromy.

start_pair_using_newton::Bool = false: if true, the algorithm tries to compute a start pair for monodromy by using Newton's methods. Can reduce the number of critical points, but is less stable.

Example

using Chambers
@var x y
f = [x^2 + y^2 - 1; x^2 + y^2 - 4];
chambers(f)

Chambers.affine_chambers — Function

affine_chambers(f::Vector{Expression})
affine_chambers(f::System)

Input a list of affine hypersurfaces 'f = [f1,...fk]'. Outputs the chambers in the complement of the hypersurface arrangement, their sign patterns, Euler characteristic and the indices of the critical points in each chamber. Accepts the same options as chambers.

Example

using Chambers
@var x y
f = [x^2 + y^2 - 1; x^2 + y^2 - 4];
affine_chambers(f)

Chambers.projective_chambers — Function

projective_chambers(C::ChambersResult)

Returns a vector of vectors. The entries of the vectors correspond to those chambers in C, which are fused in projective space.

The following code will return a vector of vectors of type Chamber:

using Chambers
@var x y
f = [x^2 + y^2 - 1; x^2 + y^2 - 4];
C = chambers(f)
p = projective_chambers(C)
map(pᵢ -> chambers(C)[pᵢ], p)

Documentation: Helper functions

Functions to call on Chamber.

Chambers.χ — Function

χ(C::Chamber)

Returns the Euler characteristic.

Chambers.μ — Function

μ(C::Chamber)

Returns the index vector.

Chambers.critical_points — Function

critical_points(C::Chamber)

Returns the critical points in C.

Chambers.is_bounded — Function

is_bounded(C::Chamber)

Returns a boolean that is true, if C is bounded. If that information was not computed, it simply returns nothing.

Functions to call on ChambersResult.

Chambers.nchambers — Function

nchambers(C::ChambersResult)

Returns the number of chambers in F.

Chambers.nbounded — Function

nbounded(C::ChambersResult)

Returns the number of bounded chambers in C.

Chambers.nunbounded — Function

nunbounded(C::ChambersResult)

Returns the number of unbounded chambers in C.

Chambers.bounded — Function

bounded(C::ChambersResult)

Returns the bounded chambers in C.

Chambers.unbounded — Function

unbounded(C::ChambersResult)

Returns the unbounded chambers in C.

Chambers.euler_characteristics — Function

euler_characteristics(C::ChambersResult)

Returns the Euler characteristics of the chambers in C.

Chambers.index_vectors — Function

euler_characteristics(C::ChambersResult)

Returns the index vectors of the chambers in C.

Chambers.ncritical_complex — Function

ncritical_complex(C::ChambersResult)

Returns the number of complex critical points of the Morse function in C.

Chambers.ncritical_real — Function

ncritical_real(C::ChambersResult)

Returns the number of real critical points of the Morse function in C.

Chambers.g — Function

g(C::ChambersResult)

Returns the Morse function in C.